% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chaves.R
\name{adicionar_chaves}
\alias{adicionar_chaves}
\title{Adicionar chaves ao projeto}
\usage{
adicionar_chaves(...)
}
\arguments{
\item{...}{Uma sequência de argumentos nomeados na forma:
\code{NOME_DA_CHAVE = "valor"}}
}
\description{
Adiciona chaves de API para serem utilizadas no projeto.
}
\section{Detalhes}{

É considerada uma boa prática nomear variáveis de ambiente com letras
maiúsculas. A função processa os nomes pra deixá-los em maiúsculas, caso
tenha sido dado em minúsculas.
}

\examples{
if (interactive()) {

  adicionar_chaves(CHAVEEXEMPLO = "valor_ficticio")

  atualizar_renviron()

}
}
