% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chaves.R
\name{atualizar_renviron}
\alias{atualizar_renviron}
\title{Atualizar .Renviron com chaves do projeto}
\usage{
atualizar_renviron()
}
\description{
Inclui as chaves do projeto no seu .Renviron. Dessa forma elas podem ser
utilizadas usando \code{Sys.getenv("chave")}. É necessário reiniciar a sessão do R
depois de atualizar pra poder usar as chaves assim.
}
\section{Chaves existentes}{

Atualmente inclui as chaves:
\itemize{
\item \code{GOOGLE1}, \code{GOOGLE2} e \code{GOOGLE3}, da API do Google Maps;
\item \code{MAPBOX}, do Mapbox.
}
}

\examples{
if (interactive()) {

  atualizar_renviron()

  # depois de reiniciar o R
  Sys.getenv("GOOGLE1")
}

}
