% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checa_wd_valido.R
\name{checa_wd_valido}
\alias{checa_wd_valido}
\title{Checar se o working directory é válido}
\usage{
checa_wd_valido()
}
\description{
Checa se o working directory é válido: se ele está dois níveis abaixo da
pasta raiz do projeto (ou seja, se você por exemplo num projeto dentro da sua
página \code{git_seunome} pessoal).
}
\keyword{internal}
