% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chaves.R
\name{remover_chaves}
\alias{remover_chaves}
\title{Remover chaves do projeto}
\usage{
remover_chaves(chaves)
}
\arguments{
\item{chaves}{Um vetor de strings com o nome das chaves a serem retiradas.
Lembre-se que as chaves estão todos com os nomes em letras maiúsculas.}
}
\description{
Remove chaves do arquivo de chaves do projeto
}
\examples{
if (interactive()) {

  remover_chaves(c("CHAVEEXEMPLO", "OUTROEXEMPLO"))

  atualizar_renviron()

}
}
